+++ coders/exr.c
@@ -206,8 +206,8 @@ static Image *ReadEXRImage(const ImageInfo *image_info,ExceptionInfo *exception)
   hdr_info=ImfInputHeader(file);
   ImfHeaderDataWindow(hdr_info,&display_window.min_x,&display_window.min_y,
     &display_window.max_x,&display_window.max_y);
-  image->columns=((size_t) display_window.max_x-display_window.min_x+1UL);
-  image->rows=((size_t) display_window.max_y-display_window.min_y+1UL);
+  image->columns=(size_t) (display_window.max_x-display_window.min_x+1L);
+  image->rows=(size_t) (display_window.max_y-display_window.min_y+1L);
   image->alpha_trait=BlendPixelTrait;
   (void) SetImageColorspace(image,RGBColorspace,exception);
   image->gamma=1.0;
@@ -253,7 +253,7 @@ static Image *ReadEXRImage(const ImageInfo *image_info,ExceptionInfo *exception)
     }
   ImfHeaderDataWindow(hdr_info,&data_window.min_x,&data_window.min_y,
     &data_window.max_x,&data_window.max_y);
-  columns=((size_t) data_window.max_x-data_window.min_x+1UL);
+  columns=(size_t) (data_window.max_x-data_window.min_x+1L);
   if ((display_window.min_x > data_window.max_x) ||
       (display_window.min_x+(int) image->columns <= data_window.min_x))
     scanline=(ImfRgba *) NULL;
@@ -292,8 +292,7 @@ static Image *ReadEXRImage(const ImageInfo *image_info,ExceptionInfo *exception)
         continue;
       }
     (void) memset(scanline,0,columns*sizeof(*scanline));
-    if (ImfInputSetFrameBuffer(file,scanline-data_window.min_x-columns*yy,1,
-      columns) == 0)
+    if (ImfInputSetFrameBuffer(file,scanline-data_window.min_x-(ssize_t) columns*yy,1,columns) == 0)
       {
         status=MagickFalse;
         break;
@@ -649,7 +648,8 @@ static MagickBooleanType WriteEXRImage(const ImageInfo *image_info,Image *image,
       scanline[x].a=half_quantum;
       p+=GetPixelChannels(image);
     }
-    ImfOutputSetFrameBuffer(file,scanline-(y*image->columns),1,image->columns);
+    ImfOutputSetFrameBuffer(file,scanline-(y*(ssize_t) image->columns),1,
+      image->columns);
     ImfOutputWritePixels(file,1);
   }
   (void) ImfCloseOutputFile(file);
